requirements.txt
google-cloud-vision
google-cloud-logging